refactoring the package name and the 2 issues pointed out by code climate
