+++ android-client/Espoo/app/src/androidTest/java/com/espoo/android/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.espoo
+package com.espoo.android
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.espoo", appContext.packageName)
+        assertEquals("com.espoo.android", appContext.packageName)
     }
 }
