+++ android-client/Espoo/app/src/main/java/com/espoo/android/model/Role.kt
@@ -1,3 +1,3 @@
-package com.example.espoo.model
+package com.espoo.android.model
 
 data class Role(var id: Int, var role_type: String)
