+++ android-client/Espoo/app/src/main/java/com/espoo/android/model/Survey.kt
@@ -1,3 +1,3 @@
-package com.example.espoo.model
+package com.espoo.android.model
 
 data class Survey(var id: Int, var name: String, var description: String, var questions_quantity: Int)
