+++ android-client/Espoo/app/src/test/java/com/espoo/android/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.espoo
+package com.espoo.android
 
 import org.junit.Test
